+++ src/lib/preferences/activity.ts
@@ -7,9 +7,9 @@ import type { Credential } from '$lib/preferences/credentials';
 import { setNewActivitiesInHome } from '$lib/homeFeedbackPreferences';
 import { invalidate } from '$app/navigation';
 import { _protectedLayoutKey } from '../../routes/[[lang]]/(protected)/+layout';
-import type { Info } from '$lib/components/organisms/scanner/tools';
 import { filesUri } from '$lib/backendUri';
 import type { Logo } from '$lib/utils/types';
+import { m } from '$lib/i18n';
 
 dayjs.extend(relativeTime);
 
@@ -75,19 +75,16 @@ export async function addActivity(activity: Activity) {
 	invalidate(_protectedLayoutKey);
 }
 
-export async function addVerificationActivity(sid: string, info: Info, success: boolean) {
+export async function addVerificationActivity(sid: string, success: boolean, verifierUrl: string | null, properties: string[] = []) {
 	const at = dayjs().unix();
-	const { asked_claims, avatar } = info;
-	const { properties } = asked_claims;
-	const propertiesArray = Object.values(properties).map((property) => property.title);
 	await addActivity({
 		type: 'verification',
-		verifier_name: info.verifier_name,
-		avatar,
+		verifier_name: verifierUrl || '',
+		avatar: { id: '', collection: '', fileName: '' },
 		success,
-		rp_name: info.rp_name,
+		rp_name: verifierUrl || '',
 		sid,
-		properties: propertiesArray,
+		properties: properties,
 		at
 	});
 }
@@ -127,12 +124,12 @@ export async function getParsedActivities(): Promise<ParsedActivity[]> {
 			parsedActivity.description = credential.description;
 			parsedActivity.credential = credential;
 			if (activity.type === 'credential') {
-				parsedActivity.message = `${credential.issuer} issued ${credential.display_name} to you`;
+				parsedActivity.message = m.issued_to_you({iss: credential.issuer, name: credential.display_name})
 			} else {
-				parsedActivity.message = `${credential.display_name} is expired`;
+				parsedActivity.message = `${credential.display_name} ${m.is_expired()}`;
 			}
 		} else if (activity.type === 'verification') {
-			const { verifier_name, success, rp_name, properties, avatar } = activity;
+			const { verifier_name, rp_name, properties, avatar } = activity;
 			parsedActivity.name = verifier_name;
 
 			if (avatar) {
@@ -141,9 +138,8 @@ export async function getParsedActivities(): Promise<ParsedActivity[]> {
 					alt_text: verifier_name
 				};
 			}
-			parsedActivity.message = `${verifier_name} verified yours: ${properties.join(', ')} via ${rp_name} and it was a ${
-				success ? 'success' : 'failure'
-			}`;
+			parsedActivity.message = m.You_send_to_verification_via_and_result_is({ properties: properties.join(','), rp_name: rp_name.split('//')[1].split('/')[0], result: activity.success ? m.successful() : m.failed() });
+			parsedActivity.description = rp_name;
 		}
 		return parsedActivity;
 	}