+++ src/routes/+layout.svelte
@@ -22,6 +22,7 @@
 	import { Network } from '@capacitor/network';
 	import { debugPopup, debugPopupContent } from '$lib/components/organisms/debug/debug';
 	import { refreshAuth } from './[[lang]]/(auth)/login/_lib';
+	import { isDark } from '$lib/isDark';
 
 	const controller = new AbortController();
 	const signal = controller.signal;
@@ -71,12 +72,8 @@
 	};
 
 	onMount(async () => {
-		const element = document.body.getElementsByTagName('ion-app')[0];
-		const styles = getComputedStyle(element);
-		// const mainColor = styles.getPropertyValue('--safe-area-background-color').trim();
-		const mainColor = styles.getPropertyValue('--ion-background-color').trim();
-		EdgeToEdge.setBackgroundColor({color: mainColor || '#000000'});
-
+		const mainColor = isDark ? '#494949' : '#b5b5b5';
+		EdgeToEdge.setBackgroundColor({ color: mainColor || '#000000' });
 
 		isConnected = (await Network.getStatus()).connected;
 		Network.addListener('networkStatusChange', async (status) => {